Allow specifying CWD
Switched processor.js over to mostly dealing w/ just absolute paths
Still a bit of relativity in output generation, but it's minor and workable
Updated browserify/rollup to work in this brave new world
Also restructured tests a bunch
