+++ package.json
@@ -45,6 +45,7 @@
   "dependencies": {
     "dependency-graph": "^0.4.1",
     "lodash.assign": "^4.0.0",
+    "lodash.defaults": "^4.0.1",
     "lodash.difference": "^4.0.0",
     "lodash.filter": "^4.2.1",
     "lodash.flatten": "^4.0.0",
