+++ test/issue-66.test.js
@@ -4,8 +4,8 @@ var assert = require("assert"),
     
     Processor = require("../src/processor");
 
-describe("modular-css", function() {
-    describe("issue 66", function() {
+describe("/issues", function() {
+    describe("/66", function() {
         it("should ignore remove calls for unknown files", function(done) {
             var processor = new Processor();
