+++ test/lib/relative.js
@@ -1,0 +1,3 @@
+"use strict";
+
+module.exports = require("../../src/_relative").bind(null, process.cwd());
