+++ test/mocha.opts
@@ -1,0 +1,3 @@
+--ui bdd
+--reporter spec
+--timeout 5000
