+++ test/processor.test.js
@@ -1,0 +1,566 @@
+"use strict";
+
+var fs      = require("fs"),
+    path    = require("path"),
+    assert  = require("assert"),
+    
+    Promise   = require("../src/_promise"),
+    Processor = require("../src/processor");
+
+// Catch unhandled promise rejections and fail the test
+process.on("unhandledRejection", function(reason) {
+    throw reason;
+});
+
+function sync(css) {
+    css.append({ selector : "a" });
+}
+
+function async(css) {
+    return new Promise(function(resolve) {
+        setTimeout(function() {
+            sync(css);
+
+            resolve();
+        }, 0);
+    });
+}
+
+describe("/processor.js", function() {
+    describe("Basics", function() {
+        it("should be a function", function() {
+            assert.equal(typeof Processor, "function");
+        });
+        
+        it("should auto-instantiate if called without new", function() {
+            /* eslint new-cap:0 */
+            assert(Processor() instanceof Processor);
+        });
+    });
+
+    describe("functionality", function() {
+        beforeEach(function() {
+            this.processor = new Processor();
+        });
+        
+        describe("options", function() {
+            describe("namer", function() {
+                it("should use a custom naming function", function(done) {
+                    var id        = path.resolve("./test/specimens/simple.css"),
+                        processor = new Processor({
+                            namer : function(filename, selector) {
+                                return filename + selector;
+                            }
+                        });
+                        
+                    processor.string("./test/specimens/simple.css", ".wooga { }").then(function(result) {
+                        var file = result.files[id];
+                    
+                        assert.deepEqual(result.exports, {
+                            wooga : id + "wooga"
+                        });
+
+                        assert.equal(typeof file, "object");
+
+                        assert.deepEqual(file.compositions, {
+                            wooga : [ id + "wooga" ]
+                        });
+
+                        assert.equal(file.text, ".wooga { }");
+                        assert.equal(file.processed.root.toResult().css, "." + id + "wooga { }");
+
+                        done();
+                    })
+                    .catch(done);
+                });
+            });
+
+            describe("before", function() {
+                it("should run sync postcss plugins before processing", function(done) {
+                    var processor = new Processor({
+                            before : [ sync ]
+                        });
+                    
+                    processor.string("test/specimens/sync-before.css", "").then(function() {
+                        return processor.output();
+                    }).then(function(result) {
+                        assert.equal(
+                            result.css,
+                            "/* test/specimens/sync-before.css */\n" +
+                            "a {}"
+                        );
+                        
+                        done();
+                    })
+                    .catch(done);
+                });
+
+                it("should run async postcss plugins before processing", function(done) {
+                    var processor = new Processor({
+                            before : [ async ]
+                        });
+                    
+                    processor.string("test/specimens/async-before.css", "").then(function() {
+                        return processor.output();
+                    }).then(function(result) {
+                        assert.equal(
+                            result.css,
+                            "/* test/specimens/async-before.css */\n" +
+                            "a {}"
+                        );
+                        
+                        done();
+                    })
+                    .catch(done);
+                });
+            });
+            
+            describe("after", function() {
+                var css =
+                        "/* test/specimens/relative.css */\n" +
+                        ".mc592b2d8f_wooga {\n" +
+                        "    color: red;\n" +
+                        "    background: url(\"./folder/to.png\")\n" +
+                        "}\n" +
+                        "a {}";
+                
+                it("should use postcss-url by default", function(done) {
+                    var processor = this.processor;
+
+                    processor.file("./test/specimens/relative.css").then(function() {
+                        return processor.output({ to : "./test/output/relative.css" });
+                    })
+                    .then(function(result) {
+                        assert.equal(
+                            result.css + "\n",
+                            fs.readFileSync("./test/results/processor/relative.css", "utf8")
+                        );
+
+                        done();
+                    })
+                    .catch(done);
+                });
+                
+                it("should run sync postcss plugins", function(done) {
+                    var processor = new Processor({
+                            after : [ sync ]
+                        });
+
+                    processor.file("./test/specimens/relative.css").then(function() {
+                        return processor.output({ to : "./test/output/relative.css" });
+                    })
+                    .then(function(result) {
+                        assert.equal(
+                            result.css,
+                            css
+                        );
+
+                        done();
+                    })
+                    .catch(done);
+                });
+                
+                it("should run async postcss plugins", function(done) {
+                    var processor = new Processor({
+                            after : [ async ]
+                        });
+
+                    processor.file("./test/specimens/relative.css").then(function() {
+                        return processor.output({ to : "./test/output/relative.css" });
+                    })
+                    .then(function(result) {
+                        assert.equal(
+                            result.css,
+                            css
+                        );
+
+                        done();
+                    })
+                    .catch(done);
+                });
+            });
+            
+            describe("done", function() {
+                it("should run sync postcss plugins done processing", function(done) {
+                    var processor = new Processor({
+                            done : [ sync ]
+                        });
+                    
+                    processor.string("test/specimens/sync-done.css", "").then(function() {
+                        return processor.output();
+                    }).then(function(result) {
+                        assert.equal(
+                            result.css,
+                            "/* test/specimens/sync-done.css */\n" +
+                            "a {}"
+                        );
+                        
+                        done();
+                    })
+                    .catch(done);
+                });
+                
+                it("should run async postcss plugins done processing", function(done) {
+                    var processor = new Processor({
+                            done : [ async ]
+                        });
+                    
+                    processor.string("test/specimens/async-done.css", "").then(function() {
+                        return processor.output();
+                    }).then(function(result) {
+                        assert.equal(
+                            result.css,
+                            "/* test/specimens/async-done.css */\n" +
+                            "a {}"
+                        );
+                        
+                        done();
+                    })
+                    .catch(done);
+                });
+            });
+            
+            describe("map", function() {
+                it("should generate source maps", function(done) {
+                    var processor = new Processor({ map : true });
+                    
+                    processor.file("./test/specimens/start.css").then(function() {
+                        return processor.output();
+                    })
+                    .then(function(result) {
+                        assert.equal(
+                            result.css + "\n",
+                            fs.readFileSync("./test/results/processor/source-map.css", "utf8")
+                        );
+                        
+                        done();
+                    })
+                    .catch(done);
+                });
+            });
+        });
+        
+        describe("Methods", function() {
+            describe(".string()", function() {
+                it("should process a string", function(done) {
+                    this.processor.string("./test/specimens/simple.css", ".wooga { }").then(function(result) {
+                        var file = result.files[path.resolve("./test/specimens/simple.css")];
+                    
+                        assert.deepEqual(result.exports, {
+                            wooga : "mc08e91a5b_wooga"
+                        });
+                        
+                        assert.equal(typeof file, "object");
+                        
+                        assert.deepEqual(file.compositions, {
+                            wooga : [ "mc08e91a5b_wooga" ]
+                        });
+                        
+                        assert.equal(file.text, ".wooga { }");
+                        assert.equal(file.processed.root.toResult().css, ".mc08e91a5b_wooga { }");
+                        
+                        done();
+                    })
+                    .catch(done);
+                });
+            });
+            
+            describe(".file()", function(done) {
+                it("should process a file", function() {
+                    this.processor.file("./test/specimens/simple.css").then(function(result) {
+                        var file = result.files[path.resolve("./test/specimens/simple.css")];
+                        
+                        assert.deepEqual(result.exports, {
+                            wooga : "mc08e91a5b_wooga"
+                        });
+                        
+                        assert.equal(typeof file, "object");
+                        
+                        assert.deepEqual(file.compositions, {
+                            wooga : [ "mc08e91a5b_wooga" ]
+                        });
+                        
+                        assert.equal(file.text, ".wooga { color: red; }\n");
+                        assert.equal(file.processed.root.toResult().css, ".mc08e91a5b_wooga { color: red; }\n");
+                        
+                        done();
+                    })
+                    .catch(done);
+                });
+            });
+            
+            describe(".remove()", function() {
+                it("should remove a file", function(done) {
+                    var processor = this.processor;
+
+                    processor.string("./test/specimens/simple.css", ".wooga { }").then(function() {
+                        processor.remove(path.resolve("./test/specimens/simple.css"));
+                        
+                        assert.deepEqual(processor.dependencies(), []);
+                        
+                        done();
+                    })
+                    .catch(done);
+                });
+                
+                it("should remove multiple files", function(done) {
+                    var processor = this.processor;
+
+                    processor.string("./test/specimens/a.css", ".aooga { }")
+                    .then(function() {
+                        return processor.string("./test/specimens/b.css", ".booga { }");
+                    })
+                    .then(function() {
+                        return processor.string("./test/specimens/c.css", ".cooga { }");
+                    })
+                    .then(function() {
+                        processor.remove([
+                            path.resolve("./test/specimens/a.css"),
+                            path.resolve("./test/specimens/b.css")
+                        ]);
+                        
+                        assert.deepEqual(processor.dependencies(), [
+                            path.resolve("./test/specimens/c.css")
+                        ]);
+
+                        done();
+                    })
+                    .catch(done);
+                });
+            });
+            
+            describe(".dependencies()", function() {
+                it("should return the dependencies of the specified file", function(done) {
+                    var processor = this.processor;
+
+                    processor.file("./test/specimens/start.css")
+                    .then(function() {
+                        assert.deepEqual(
+                            processor.dependencies(path.resolve("./test/specimens/start.css")),
+                            [
+                                path.resolve("./test/specimens/folder/folder.css"),
+                                path.resolve("./test/specimens/local.css")
+                            ]
+                        );
+
+                        done();
+                    })
+                    .catch(done);
+                });
+                
+                it("should return the overall order of dependencies if no file is specified", function(done) {
+                    var processor = this.processor;
+
+                    processor.file("./test/specimens/start.css")
+                    .then(function() {
+                        assert.deepEqual(processor.dependencies(), [
+                            path.resolve("./test/specimens/folder/folder.css"),
+                            path.resolve("./test/specimens/local.css"),
+                            path.resolve("./test/specimens/start.css")
+                        ]);
+                        
+                        done();
+                    })
+                    .catch(done);
+                });
+            });
+            
+            describe(".output()", function() {
+                it("should return a postcss result", function(done) {
+                    var processor = this.processor;
+
+                    processor.file("./test/specimens/start.css").then(function() {
+                        return processor.output();
+                    })
+                    .then(function(result) {
+                        assert.equal(
+                            result.css + "\n",
+                            fs.readFileSync("./test/results/processor/start.css", "utf8")
+                        );
+
+                        done();
+                    })
+                    .catch(done);
+                });
+                
+                it("should generate css representing the output from all added files", function(done) {
+                    var processor = this.processor;
+
+                    processor.file("./test/specimens/start.css").then(function() {
+                        return processor.file("./test/specimens/simple.css");
+                    })
+                    .then(function() {
+                        return processor.output();
+                    })
+                    .then(function(result) {
+                        assert.equal(
+                            result.css + "\n",
+                            fs.readFileSync("./test/results/processor/output-all.css", "utf8")
+                        );
+
+                        done();
+                    })
+                    .catch(done);
+                });
+
+                it("should avoid duplicating files in the output", function(done) {
+                    var processor = this.processor;
+
+                    processor.file("./test/specimens/start.css").then(function() {
+                        return processor.file("./test/specimens/local.css");
+                    })
+                    .then(function() {
+                        return processor.output();
+                    })
+                    .then(function(result) {
+                        assert.equal(
+                            result.css + "\n",
+                            fs.readFileSync("./test/results/processor/avoid-duplicates.css", "utf8")
+                        );
+
+                        done();
+                    })
+                    .catch(done);
+                });
+            });
+        });
+        
+        it("should export an object of space-separated strings", function(done) {
+            this.processor.string(
+                "./test/specimens/simple.css",
+                ".fooga { color: red; } .booga { background: #000; } .tooga { composes: fooga, booga; }"
+            ).then(function(result) {
+                assert.deepEqual(result.exports, {
+                    fooga : "mc08e91a5b_fooga",
+                    booga : "mc08e91a5b_booga",
+                    tooga : "mc08e91a5b_fooga mc08e91a5b_booga"
+                });
+                
+                done();
+            })
+            .catch(done);
+        });
+
+        it("should scope classes, ids, and keyframes", function(done) {
+            this.processor.string(
+                "./test/specimens/simple.css",
+                "@keyframes kooga { } #fooga { } .wooga { }"
+            ).then(function(result) {
+                var file = result.files[path.resolve("./test/specimens/simple.css")];
+                
+                assert.deepEqual(result.exports, {
+                    fooga : "mc08e91a5b_fooga",
+                    kooga : "mc08e91a5b_kooga",
+                    wooga : "mc08e91a5b_wooga"
+                });
+
+                assert.equal(typeof file, "object");
+
+                assert.deepEqual(file.compositions, {
+                    fooga : [ "mc08e91a5b_fooga" ],
+                    kooga : [ "mc08e91a5b_kooga" ],
+                    wooga : [ "mc08e91a5b_wooga" ]
+                });
+
+                assert.equal(file.text, "@keyframes kooga { } #fooga { } .wooga { }");
+                assert.equal(
+                    file.processed.root.toResult().css,
+                    "@keyframes mc08e91a5b_kooga { } " +
+                    "#mc08e91a5b_fooga { } " +
+                    ".mc08e91a5b_wooga { }"
+                );
+
+                done();
+            })
+            .catch(done);
+        });
+
+        it("should export identifiers and their classes", function(done) {
+            this.processor.file("./test/specimens/start.css").then(function(result) {
+                var file = result.files[path.resolve("./test/specimens/start.css")];
+            
+                assert.deepEqual(result.exports, {
+                    wooga : "mc04cb4cb2_booga",
+                    booga : "mc61f0515a_booga",
+                    tooga : "mc61f0515a_tooga"
+                });
+
+                assert.equal(file.text, fs.readFileSync("./test/specimens/start.css", "utf8"));
+                assert.equal(
+                    file.processed.root.toResult().css,
+                    ".mc61f0515a_booga { color: red; background: blue; }\n" +
+                    ".mc61f0515a_tooga { border: 1px solid white; }\n"
+                );
+
+                assert.deepEqual(file.values, {
+                    folder : "white",
+                    one    : "red",
+                    two    : "blue"
+                });
+
+                assert.deepEqual(file.compositions, {
+                    wooga : [ "mc04cb4cb2_booga" ],
+                    booga : [ "mc61f0515a_booga" ],
+                    tooga : [ "mc61f0515a_tooga" ]
+                });
+
+                file = result.files[path.resolve("./test/specimens/local.css")];
+
+                assert.equal(file.text, fs.readFileSync("./test/specimens/local.css", "utf8"));
+                assert.equal(file.processed.root.toResult().css, ".mc04cb4cb2_booga { background: green; }\n");
+
+                assert.deepEqual(file.values, {
+                    folder : "white",
+                    one    : "red",
+                    two    : "blue"
+                });
+
+                assert.deepEqual(file.compositions, {
+                    booga : [ "mc04cb4cb2_booga" ],
+                    looga : [ "mc04cb4cb2_booga" ]
+                });
+
+                file = result.files[path.resolve("./test/specimens/folder/folder.css")];
+
+                assert.equal(file.text, fs.readFileSync("./test/specimens/folder/folder.css", "utf8"));
+                assert.equal(file.processed.root.toResult().css, ".mc04bb002b_folder { margin: 2px; }\n");
+
+                assert.deepEqual(file.values, {
+                    folder : "white"
+                });
+                
+                assert.deepEqual(file.compositions, {
+                    folder : [ "mc04bb002b_folder" ]
+                });
+
+                done();
+            });
+        });
+        
+        describe("bad imports", function() {
+            var invalid = "Unable to locate \"../local.css\" from \"" + path.resolve("invalid") + "\"";
+            
+            it("should fail if a value imports a non-existant reference", function(done) {
+                this.processor.string(
+                    "./invalid/value.css",
+                    "@value not-real from \"../local.css\";"
+                ).catch(function(error) {
+                    assert.equal(error.message, invalid);
+
+                    done();
+                });
+            });
+            
+            it("should fail if a composition imports a non-existant reference", function(done) {
+                this.processor.string(
+                    "./invalid/composition.css",
+                    ".wooga { composes: fake from \"../local.css\"; }"
+                ).catch(function(error) {
+                    assert.equal(error.message, invalid);
+
+                    done();
+                });
+            });
+        });
+    });
+});
