+++ test/results/browserify/avoid-duplicates-local.js
