--- test/results/browserify/export-all-identifiers.json
@@ -1,6 +1,0 @@
-{
-    "test/specimens/multiple.css": {
-        "fooga": "mcdc9e477f_fooga",
-        "wooga": "mcdc9e477f_fooga mcdc9e477f_wooga"
-    }
-}
