+++ test/results/browserify/export-identifiers.js
